+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade handlebars to 4.7.6
 - Force fallback method for dragging items (fix mobile support)
 - Fix new task dialog display glitches
+- Improve check list item edition
 
 ### Removed